+++ src/app/_components/tableLayout.tsx
@@ -1,4 +1,5 @@
 'use client'
+import { LocationUrlContext } from "@/_contexts/context";
 import { checkSamePathName, checkTabPathname } from "@/_libs/location.utils";
 import CardContent from "@mui/material/CardContent";
 import Paper from "@mui/material/Paper";
@@ -8,37 +9,10 @@ import Tabs from "@mui/material/Tabs";
 import { useSession } from "next-auth/react";
 import { redirect, useParams, usePathname, useRouter, useSearchParams } from "next/navigation";
 import {
-  createContext, Dispatch, ReactNode,
-  RefObject, SetStateAction,
-  SyntheticEvent, useContext,
+  ReactNode,
+  SyntheticEvent,
   useEffect, useRef, useState
 } from "react";
-
-type LocationStateContext = {
-  selected: RefObject<string>;
-  locationId: number;
-  setLocationId: Dispatch<SetStateAction<number>>;
-}
-
-export type TRemarkAssetCountLine = {
-  id: string;
-  remark: string;
-}
-
-export type ReportStateContext = {
-  DocumentNumber: number | undefined;
-  update: boolean;
-  remark?: TRemarkAssetCountLine[];
-  setDocumentNumber: Dispatch<SetStateAction<number | undefined>>;
-  setRefetchReport: Dispatch<SetStateAction<boolean>>;
-  setUpdate: Dispatch<SetStateAction<boolean>>;
-  setRemark?: Dispatch<SetStateAction<TRemarkAssetCountLine[]>>;
-  setSearch:  Dispatch<SetStateAction<boolean>>;
-}
-
-//31 line of console.log
-export const LocationUrlContext = createContext<LocationStateContext | null>(null);
-export const ReportContext = createContext<ReportStateContext | null>(null);
 
 export default function TableLayout({
   children
@@ -107,20 +81,4 @@ export default function TableLayout({
     </CardContent>
     {/* </Card > */}
   </>
-}
-
-export function useLocationUrlContext() {
-  const context = useContext(LocationUrlContext)
-  if (!context) {
-    throw new Error("useLocatoinUrlContext must be use within Context provider")
-  }
-  return context
-}
-
-export function useReportContext() {
-  const context = useContext(ReportContext)
-  if (!context) {
-    throw new Error("useReportContext must be use within Context provider")
-  }
-  return context
 }
