+++ src/app/_libs/location.utils.ts
@@ -1,4 +1,4 @@
-import { TLocation } from "@/_types/snipe-it.type";
+import { TLocation } from "../_types/snipe-it.type";
 import { createGateway } from "@/api/next.api";
 
 export function getParentLocation(data: TLocation[]) : TLocation[] {
