+++ src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginComponent from "@/_components/loginComponent";
+import LoginComponent from "@/auth/login/loginComponent";
 
 export default async function LoginPage() {
   return (
