+++ src/app/auth/register/registerComponent.tsx
