+++ src/app/forbidden/page.tsx
@@ -1,4 +1,4 @@
-import ForbiddenComponent from "@/_components/forbiddenComponent";
+import ForbiddenComponent from "@/forbidden/forbiddenComponent";
 
 export default async function Forbidden() {
     return (
