+++ src/app/reports/[reportId]/page.tsx
@@ -1,7 +1,7 @@
-import AssetReport from '@/_components/assetReport'
+import AssetReport from '@/reports/[reportId]/assetReport'
 import { getAssetCountLineByAssetCount, getAssetCountReport } from '@/_libs/report.utils'
-import { AssetCountWithLineAndLocation } from '@/_types/interfaces'
-import { AssetCountLine, User } from '@/_types/types'
+import { AssetCountWithLineAndLocation } from '../../_types/interfaces'
+import { AssetCountLine, User } from '../../_types/types'
 import { GetAllUserPrisma } from '@/api/report.api'
 import { fetchLocations } from '@/api/snipe-it/snipe-it.api'
 import { getSession } from 'auth'
