+++ src/app/reports/page.tsx
@@ -1,6 +1,6 @@
 'use server'
 import { fetchLocations } from "@/api/snipe-it/snipe-it.api"
-import ReportComponent from "@/_components/reportComponent";
+import ReportComponent from "@/reports/_components/reportComponent";
 import { getChildrenLocation, getOtherLocation, getParentFromChildId, getParentLocation } from '@/_libs/location.utils';
 import { TLocation } from "@/_types/snipe-it.type";
 import { getSession } from "auth";
